Move iterator tests and add comments
