+++ puzzled_common/src/shape/iterators.rs
@@ -3,34 +3,38 @@ use ndarray::iter::{IndexedIter, Iter};
 use ndarray::Ix2;
 
 impl Shape {
+    /// Iterates over the values of the shape.
     pub fn iter(&self) -> Iter<'_, bool, Ix2> {
         self.data.iter()
     }
 
+    /// Iterates over the values of the shape while also being given the index of the value.
     pub fn indexed_iter(&self) -> IndexedIter<'_, bool, Ix2> {
         self.data.indexed_iter()
     }
 
-    pub fn rotations_flips_iter(&self) -> TileRotationIterator {
-        TileRotationIterator::new(self.clone())
+    /// Iterates over all rotations of the shape.
+    /// Duplicates are not removed.
+    pub fn rotations_flips_iter(&self) -> ShapeRotationIterator {
+        ShapeRotationIterator::new(self.clone())
     }
 }
 
-pub struct TileRotationIterator {
+pub struct ShapeRotationIterator {
     current: Shape,
     iteration: u8,
 }
 
-impl TileRotationIterator {
-    fn new(tile: Shape) -> Self {
+impl ShapeRotationIterator {
+    fn new(shape: Shape) -> Self {
         Self {
-            current: tile,
+            current: shape,
             iteration: 0,
         }
     }
 }
 
-impl Iterator for TileRotationIterator {
+impl Iterator for ShapeRotationIterator {
     type Item = Shape;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -44,5 +48,51 @@ impl Iterator for TileRotationIterator {
         self.current.rotate_counterclockwise();
         self.iteration += 1;
         Some(current)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::shape::shape_square;
+
+    #[test]
+    fn test_tile_rotation_iterator() {
+        let base = shape_square(&[[true, false], [false, false]]);
+        let mut iter = base.rotations_flips_iter();
+
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[true, false], [false, false]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, false], [true, false]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, false], [false, true]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, true], [false, false]]))
+        );
+
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[true, false], [false, false]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, false], [true, false]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, false], [false, true]]))
+        );
+        assert_eq!(
+            iter.next(),
+            Some(shape_square(&[[false, true], [false, false]]))
+        );
+        assert_eq!(iter.next(), None);
     }
 }
